+++ src/syft/lib/python/primitive_factory.py
@@ -15,6 +15,7 @@ from ...logger import traceback_and_raise
 from .primitive_interface import PyPrimitive
 
 NoneType = type(None)
+NotImplementedType = type(NotImplemented)
 
 primitives = [
     bool,
@@ -28,6 +29,7 @@ primitives = [
     slice,
     None,
     NoneType,
+    NotImplementedType,
     str,
     UserDict,
     UserList,
@@ -47,6 +49,7 @@ PrimitiveType = Union[
     slice,
     None,
     NoneType,
+    NotImplementedType,
     str,
     UserDict,
     UserList,
@@ -67,7 +70,7 @@ class PrimitiveFactory(ABC):
 
     @staticmethod
     def generate_primitive(
-        value: Union[PrimitiveType, type(NotImplemented), PyPrimitive],  # type: ignore
+        value: Union[PrimitiveType, PyPrimitive],  # type: ignore
         id: Optional[UID] = None,
         recurse: bool = False,
     ) -> Any:
